requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
sniffio==1.3.1
Flask==2.3.3
Flask-Cors==3.0.10
proxyapi==1.0.2
openai==0.27.8
python-dotenv==1.0.0

